+++ demo/demoproject/download/urls.py
@@ -1,4 +1,4 @@
-"""URLconf for tests."""
+"""URL mapping."""
 from django.conf.urls import patterns, include, url
 
 
@@ -7,6 +7,4 @@ urlpatterns = patterns('demoproject.download.views',
         name='download_hello_world'),
     url(r'^document/(?P<slug>[a-zA-Z0-9_-]+)/$', 'download_document',
         name='download_document'),
-    url(r'^document-nginx/(?P<slug>[a-zA-Z0-9_-]+)/$',
-        'download_document_nginx', name='download_document_nginx'),
 )
