+++ demo/demoproject/download/views.py
@@ -1,7 +1,6 @@
 from os.path import abspath, dirname, join
 
 from django_downloadview import DownloadView, ObjectDownloadView
-from django_downloadview.nginx import x_accel_redirect
 
 from demoproject.download.models import Document
 
@@ -15,7 +14,3 @@ download_hello_world = DownloadView.as_view(filename=hello_world_file,
                                             storage=None)
 
 download_document = ObjectDownloadView.as_view(model=Document)
-
-download_document_nginx = x_accel_redirect(download_document,
-                                           media_root='/var/www/files',
-                                           media_url='/download-optimized')
