+++ demo/demoproject/nginx/__init__.py
@@ -1,0 +1,1 @@
+"""Nginx optimizations applied to demoproject.download."""
