+++ docs/optimizations/nginx.txt
@@ -15,70 +15,104 @@ See `Nginx X-accel documentation`_ for details.
 Configure some download view
 ****************************
 
-As an example, let's consider an application called "myapp".
+Let's start in the situation described in the :doc:`demo application </demo>`:
 
-:file:`settings.py`:
+* a project "demoproject"
+* an application "demoproject.download"
+* a :py:class:`django_downloadview.views.ObjectDownloadView` view serves files
+  of a "Document" model.
+
+We are to make it more efficient with Nginx.
+
+.. note::
+
+   Examples below are taken from the :doc:`demo project </demo>`.
+
+
+***********
+Write tests
+***********
+
+Use :py:func:`django_downloadview.nginx.assert_x_accel_redirect` function as
+a shortcut in your tests.
+
+:file:`demo/demoproject/nginx/tests.py`:
+
+.. literalinclude:: ../../demo/demoproject/nginx/tests.py
+   :language: python
+   :emphasize-lines: 5, 25-34
+
+Right now, this test should fail, since you haven't implemented the view yet.
+
+
+************
+Setup Django
+************
+
+At the end of this setup, the test should pass, but you still have to `setup
+Nginx`_!
+
+You have two options: global setup with a middleware, or per-view setup with
+decorators. 
+
+Global delegation, with XAccelRedirectMiddleware
+================================================
+
+If you want to delegate all file downloads to Nginx, then use
+:py:class:`django_downloadview.nginx.XAccelRedirectMiddleware`.
+
+Register it in your settings:
 
 .. code-block:: python
 
-   INSTALLED_APPS = (
+   MIDDLEWARE_CLASSES = (
        # ...
-       'myapp',
+       'django_downloadview.nginx.XAccelRedirectMiddleware',
        # ...
    )
-   MYAPP_STORAGE_LOCATION = '/var/www/files/'  # Could be MEDIA_ROOT for public
-                                               # files.
 
-This application holds a ``Document`` model.
-
-:file:`myapp/models.py`:
+Setup the middleware:
 
 .. code-block:: python
 
-   from django.conf import settings
-   from django.core.files.storage import FileSystemStorage
-   from django.db import models
+   NGINX_DOWNLOAD_MIDDLEWARE_MEDIA_ROOT = MEDIA_ROOT  # Could be elsewhere.
+   NGINX_DOWNLOAD_MIDDLEWARE_MEDIA_URL = '/proxied-download'
 
-
-   storage = FileSystemStorage(location=settings.MYAPP_STORAGE_LOCATION)
-
-
-   class Document(models.Model):
-       file = models.FileField(storage=storage, upload_to='document')
-
-Notice the ``storage`` and ``upload_to`` parameters: files for ``Document``
-model live in :file:`/var/www/files/document/` folder.
-
-Then we configured a download view for this model, restricted to authenticated
-users:
-
-:file:`myapp/urls.py`:
+Optionally fine-tune the middleware. Default values are ``None``, which means
+"use Nginx's defaults".
 
 .. code-block:: python
 
-   from django.conf.urls import url, url_patterns
-   from django.contrib.auth.decorators import login_required
+   NGINX_DOWNLOAD_MIDDLEWARE_EXPIRES = False  # Force no expiration.
+   NGINX_DOWNLOAD_MIDDLEWARE_WITH_BUFFERING = False  # Force buffering off.
+   NGINX_DOWNLOAD_MIDDLEWARE_LIMIT_RATE = False  # Force limit rate off.
 
-   from django_downloadview import ObjectDownloadView
+Local delegation, with x_accel_redirect decorator
+=================================================
 
-   from myapp.models import Document
+If you want to delegate file downloads to Nginx on a per-view basis, then use
+:py:func:`django_downloadview.nginx.x_accel_redirect` decorator.
+
+:file:`demo/demoproject/nginx/views.py`:
+
+.. literalinclude:: ../../demo/demoproject/nginx/views.py
+   :language: python
+
+And use it in som URL conf, as an example in
+:file:`demo/demoproject/nginx/urls.py`:
+
+.. literalinclude:: ../../demo/demoproject/nginx/urls.py
+   :language: python
+
+.. note::
+
+   In real life, you'd certainly want to replace the "download_document" view
+   instead of registering a new view.
 
 
-   download = login_required(ObjectDownloadView.as_view(model=Document))
-
-   url_patterns = ('',
-       url('^document/(?P<pk>[0-9]+/download/$', download, name='download'),
-   )
-
-As is, Django is to serve the files, i.e. load chunks into memory and stream
-them.
-
-Nginx is much more efficient for the actual streaming... Let's use it!
-
-
-***************
-Configure Nginx
-***************
+***********
+Setup Nginx
+***********
 
 See `Nginx X-accel documentation`_ for details.
 
@@ -103,7 +137,7 @@ Here is what you could have in :file:`/etc/nginx/sites-available/default`:
        #
        # See http://wiki.nginx.org/X-accel
        # and https://github.com/benoitbryon/django-downloadview
-       location /optimized-download {
+       location /proxied-download {
            internal;
            # Location to files on disk.
            # See Django's settings.NGINX_DOWNLOAD_MIDDLEWARE_MEDIA_ROOT
@@ -124,76 +158,12 @@ section.
 
 .. note::
 
-   ``/optimized-download`` is not available for the client, i.e. users
+   ``/proxied-download`` is not available for the client, i.e. users
    won't be able to download files via ``/optimized-download/<filename>``.
 
 .. warning::
 
    Make sure Nginx can read the files to download! Check permissions.
-
-
-************************************************
-Global delegation, with XAccelRedirectMiddleware
-************************************************
-
-If you want to delegate all file downloads to Nginx, then use
-:py:class:`django_downloadview.nginx.XAccelRedirectMiddleware`.
-
-Register it in your settings:
-
-.. code-block:: python
-
-   MIDDLEWARE_CLASSES = (
-       # ...
-       'django_downloadview.nginx.XAccelRedirectMiddleware',
-       # ...
-   )
-
-Setup the middleware:
-
-.. code-block:: python
-
-   NGINX_DOWNLOAD_MIDDLEWARE_MEDIA_ROOT = MYAPP_STORAGE_LOCATION
-   NGINX_DOWNLOAD_MIDDLEWARE_MEDIA_URL = '/optimized-download'
-
-Optionally fine-tune the middleware. Default values are ``None``, which means
-"use Nginx's defaults".
-
-.. code-block:: python
-
-   NGINX_DOWNLOAD_MIDDLEWARE_EXPIRES = False  # Force no expiration.
-   NGINX_DOWNLOAD_MIDDLEWARE_WITH_BUFFERING = False  # Force buffering off.
-   NGINX_DOWNLOAD_MIDDLEWARE_LIMIT_RATE = False  # Force limit rate off.
-
-
-*************************************************
-Local delegation, with x_accel_redirect decorator
-*************************************************
-
-If you want to delegate file downloads to Nginx on a per-view basis, then use
-:py:func:`django_downloadview.nginx.x_accel_redirect` decorator.
-
-Adapt :file:`myapp/urls.py`:
-
-.. code-block:: diff
-
-   from django.conf.urls import url, url_patterns
-   from django.contrib.auth.decorators import login_required
-
-   from django_downloadview import ObjectDownloadView
-   + from django_downloadview.nginx import x_accel_redirect
-
-   from myapp.models import Document
-
-
-   download = login_required(ObjectDownloadView.as_view(model=Document))
-   + download = x_accel_redirect(download,
-   +                             media_root=settings.MY_APP_STORAGE_LOCATION,
-   +                             media_url='/optimized-download')
-
-   url_patterns = ('',
-       url('^document/(?P<pk>[0-9]+/download/$', download, name='download'),
-   )
 
 
 *************
